requirements-phase2.txt
# Phase 2 Additional Requirements
# Install with: pip install -r requirements-phase2.txt

# ============================================================================
# AI Agent Framework Enhancement
# ============================================================================
langgraph>=0.2.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0
langsmith>=0.1.0

# ============================================================================
# Production & Monitoring
# ============================================================================
# Redis - TODO: Add in future for distributed state & token blacklist
# redis>=4.6.0

# Rate limiting
Flask-Limiter>=3.5.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Security
bcrypt>=4.1.0
cryptography>=41.0.0

# API validation
marshmallow>=3.20.0

# ============================================================================
# PLM Integration Foundation
# ============================================================================
# REST API clients
httpx>=0.25.0  # Async HTTP client
aiohttp>=3.9.0

# OAuth & Enterprise SSO
authlib>=1.3.0
python-jose>=3.3.0

# Data transformation
jsonpath-ng>=1.6.0
xmlschema>=3.0.0

# SAP OData
pyodata>=1.2.0

# ============================================================================
# Advanced Visualization
# ============================================================================
# WebSocket support
websockets>=12.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Data export enhancements
openpyxl>=3.1.0  # Excel
reportlab>=4.0.0  # PDF
rdflib>=7.0.0  # RDF/OSLC

# ============================================================================
# Testing & Quality
# ============================================================================
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
locust>=2.20.0  # Load testing

# Code quality
pylint>=3.0.0
bandit>=1.7.0  # Security linting

# ============================================================================
# Database & Caching
# ============================================================================
alembic>=1.13.0  # Database migrations (if needed)

# ============================================================================
# Utilities
# ============================================================================
python-multipart>=0.0.6
pytz>=2023.3
